+++ grader/checks/pylint_check.py
@@ -1,5 +1,4 @@
-"""
-Module containing the pylint check.
+"""Module containing the pylint check.
 It uses the pylint python library directly to run the check.
 """
 
@@ -7,6 +6,7 @@ import logging
 import os
 import re
 from io import StringIO
+from typing import Optional
 
 from pylint.reporters.text import TextReporter
 
@@ -25,9 +25,18 @@ class PylintCheck(ScoredCheck):
     The Pylint check class.
     """
 
-    def __init__(self, name: str, project_root: str, max_points: int, is_venv_required: bool):
-        super().__init__(name, max_points, project_root, is_venv_required)
+    def __init__(
+        self,
+        name: str,
+        project_root: str,
+        max_points: int,
+        is_venv_required: bool,
+        pylintrc_path: Optional[str] = None,
+        env_vars: Optional[dict[str, str]] = None,
+    ):
+        super().__init__(name, max_points, project_root, is_venv_required, env_vars)
         self.__pylint_max_score = 10
+        self.__pylintrc_path = pylintrc_path or const.PYLINTRC
 
     def run(self) -> ScoredCheckResult:
         """
@@ -49,13 +58,12 @@ class PylintCheck(ScoredCheck):
         logger.debug("Running pylint check on files: %s", pylint_args)
         pylint_args.append("--fail-under=0")
 
-        pylintrc_path = const.PYLINTRC
-        if os.path.exists(pylintrc_path):
-            pylint_args.extend(["--rcfile", pylintrc_path])
+        if os.path.exists(self.__pylintrc_path):
+            pylint_args.extend(["--rcfile", self.__pylintrc_path])
 
         command = [const.PYLINT_PATH] + pylint_args  # Current working directory is set in the process.run method
         try:
-            results = process.run(command, current_directory=self._project_root)
+            results = process.run(command, current_directory=self._project_root, env_vars=self.env_vars)
         except (OSError, ValueError) as error:
             logger.error("Error while running pylint: %s", error)
             raise CheckError("Error while running pylint") from error