+++ grader/checks/requirements_check.py
@@ -7,7 +7,8 @@ import logging
 import os
 
 from grader.checks.abstract_check import ScoredCheck, ScoredCheckResult
-from grader.utils.constants import REQUIREMENTS_FILENAME
+from grader.utils.constants import REQUIREMENTS_FILENAME, PYPROJECT_FILENAME
+from typing import Optional
 
 logger = logging.getLogger("grader")
 
@@ -17,10 +18,18 @@ class RequirementsCheck(ScoredCheck):
     The requirements.txt check class.
     """
 
-    def __init__(self, name: str, project_root: str, max_points: int, is_venv_required: bool):
-        super().__init__(name, max_points, project_root, is_venv_required)
+    def __init__(
+        self,
+        name: str,
+        project_root: str,
+        max_points: int,
+        is_venv_required: bool,
+        env_vars: Optional[dict[str, str]] = None,
+    ):
+        super().__init__(name, max_points, project_root, is_venv_required, env_vars)
 
         self.__requirements_path = os.path.join(self._project_root, REQUIREMENTS_FILENAME)
+        self.__pyproject_path = os.path.join(self._project_root, PYPROJECT_FILENAME)
 
     def run(self) -> ScoredCheckResult:
         """
@@ -31,6 +40,10 @@ class RequirementsCheck(ScoredCheck):
         """
         self._pre_run()
 
-        score = int(os.path.exists(self.__requirements_path)) * self.max_points
+        files_to_search = [self.__requirements_path, self.__pyproject_path]
+
+        is_one_of_files_present = any(os.path.exists(file_path) for file_path in files_to_search)
+
+        score = int(is_one_of_files_present) * self.max_points
 
         return ScoredCheckResult(self.name, score, self.max_points)